expense-auditor/requirements.txt
fastapi
uvicorn
sqlalchemy
celery
redis
python-multipart
langgraph
langchain-openai
langchain-huggingface
langchain-chroma
sentence-transformers
pydantic
python-dotenv